requirements.txt
flask
opencv-python
torch
torchvision
Pillow
